+++ how-to/workspace-platform-starter/client/src/modules/platform-override/wps-platform-override/shapes.ts
@@ -1,0 +1,7 @@
+/**
+ * Options for the wps platform override.
+ */
+export interface WpsPlatformOverrideOptions {
+	/** The name to use when logging out messages */
+	loggerName?: string;
+}